+++ packages/frontend/tsconfig.json
@@ -36,7 +36,6 @@
             ],
             "@ubichill/engine": ["../engine/src/index.ts"],
             "@ubichill/sandbox": ["../sandbox/src/index.ts"],
-            "@ubichill/sandbox/host": ["../sandbox/src/host/index.ts"],
             "@ubichill/react": ["../react/src/index.ts"],
             "@ubichill/shared": ["../shared/src/index.ts"]
         },
