+++ packages/sandbox/package.json
@@ -5,13 +5,11 @@
     "main": "src/index.ts",
     "types": "src/index.ts",
     "exports": {
-        ".": "./src/index.ts",
-        "./host": "./src/host/index.ts",
-        "./guest": "./src/guest/index.ts",
-        "./guest/jsx-runtime": "./src/guest/jsx-runtime.ts"
+        ".": "./src/index.ts"
     },
     "peerDependencies": {
         "@ubichill/engine": "workspace:*",
+        "@ubichill/sdk": "workspace:*",
         "@ubichill/shared": "workspace:*"
     },
     "devDependencies": {
