+++ packages/sandbox/src/InputCollector.ts
