--- packages/sandbox/src/guest/index.ts
@@ -1,2 +1,0 @@
-export type { EntityState, EntityStateFor, OmitId, UiRenderCostStat } from './UbiSDK';
-export { UbiSDK } from './UbiSDK';
