--- packages/sandbox/src/host/index.ts
@@ -1,35 +1,0 @@
-export {
-    createPluginFetchHandler,
-    DEFAULT_ALLOWED_DOMAINS,
-    DEMO_ALLOWED_DOMAINS,
-    fetchDirect,
-    isUrlAllowed,
-    PRODUCTION_ALLOWED_DOMAINS,
-} from './fetchHandler';
-export type {
-    FetchOptions,
-    FetchResult,
-    HostHandlers,
-    PluginHostManagerOptions,
-    PluginWorkerInfo,
-} from './PluginHostManager';
-export {
-    CAPABILITY_COMMANDS,
-    PluginHostManager,
-} from './PluginHostManager';
-export type {
-    DiagnosticCode,
-    DiagnosticHandler,
-    DiagnosticLevel,
-    MetricHandler,
-    PluginDiagnostic,
-    TickMetric,
-} from './pluginDiagnostics';
-export {
-    clearMetricHandler,
-    isMetricEnabled,
-    resetDiagnosticHandler,
-    setDiagnosticHandler,
-    setMetricHandler,
-} from './pluginDiagnostics';
-export { renderVNode } from './VNodeRenderer';
