+++ packages/sdk/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@ubichill/sdk",
     "version": "1.0.0",
-    "description": "Ubichill Plugin SDK — public type definitions for plugin developers.",
+    "description": "Ubichill Plugin SDK — UbiSDK runtime and public API for plugin developers.",
     "sideEffects": false,
     "main": "src/index.ts",
     "types": "src/index.ts",
@@ -14,7 +14,6 @@
     "peerDependencies": {
         "@ubichill/engine": "workspace:*",
         "@ubichill/react": "workspace:*",
-        "@ubichill/sandbox": "workspace:*",
         "@ubichill/shared": "workspace:*",
         "react": "^19.0.0"
     },
