+++ packages/sdk/src/ui/VNodeRenderer.ts
@@ -1,5 +1,5 @@
 /**
- * VNodeRenderer の実装は @ubichill/sandbox/host に置いてある。
+ * VNodeRenderer の実装は @ubichill/sandbox に置いてある。
  * @ubichill/react が sandbox に依存しているため、ここでは re-export のみ。
  */
-export { renderVNode } from '@ubichill/sandbox/host';
+export { renderVNode } from '@ubichill/sandbox';
